notificationService/requirements.txt
fastapi[all] 
pydantic[email]
python-dotenv
sqlalchemy
psycopg2
alembic
uvicorn
bcrypt==3.2.0 
passlib[bcrypt]
python-jose[cryptography]
twilio
